fix boostrap daemon

+++ scripts/deploy-bootstrapper.sh
@@ -8,7 +8,7 @@ log "> Deploying bootstrap node"
 log "Stopping lotus daemon"
 
 sudo systemctl stop lotus-fountain &
-sudo systemctl stop lotus-bootstrap-daemon &
+sudo systemctl stop lotus-daemon &
 sudo systemctl stop lotus-genesis-miner &
 sudo systemctl stop lotus-genesis-daemon &
 wait 
@@ -17,8 +17,10 @@ log 'Initializing repo'
 
 sudo mkdir -p /data/lotus/dev/.lotus
 sudo mkdir -p /var/log/lotus
-sudo rm -rf /root/.lotus
-sudo ln -s /data/lotus/dev/.lotus /root/.lotus
+if sudo test "-f /root/.lotus"; then
+    sudo rm -rf /root/.lotus
+    sudo ln -s /data/lotus/dev/.lotus /root/.lotus
+fi
 
 sudo cp -f lotus /usr/local/bin
 sudo cp -f lotus-miner /usr/local/bin
@@ -26,7 +28,7 @@ sudo cp -f lotus-fountain /usr/local/bin
 
 sudo cp -f scripts/lotus-genesis-daemon.service /etc/systemd/system/lotus-genesis-daemon.service
 sudo cp -f scripts/lotus-genesis-miner.service /etc/systemd/system/lotus-genesis-miner.service
-sudo cp -f scripts/lotus-bootstrap-daemon.service /etc/systemd/system/lotus-bootstrap-daemon.service
+sudo cp -f scripts/lotus-daemon.service /etc/systemd/system/lotus-daemon.service
 sudo cp -f scripts/lotus-fountain.service /etc/systemd/system/lotus-fountain.service
 
 sudo systemctl daemon-reload
@@ -38,12 +40,12 @@ sleep 30
 sudo systemctl enable lotus-genesis-miner
 sudo systemctl start lotus-genesis-miner
 
-sudo systemctl enable lotus-bootstrap-daemon
-sudo systemctl start lotus-bootstrap-daemon
+sudo systemctl enable lotus-daemon
+sudo systemctl start lotus-daemon
 
 sudo cp scripts/bootstrap.toml /root/.lotus/config.toml
 sudo bash -c "echo -e '[Metrics]\nNickname=\"Boot-bootstrap\"' >> /root/.lotus/config.toml"
-sudo systemctl restart lotus-bootstrap-daemon
+sudo systemctl restart lotus-daemon
 
 sleep 30
 
@@ -65,7 +67,7 @@ sudo systemctl start lotus-fountain
 sudo ps axu|grep "lotus"
 
 echo "fountain log:         tail -f /var/log/lotus/fountain.log"
-echo "bootstrap daemon log: tail -f /var/log/lotus/bootstrap-daemon.log"
+echo "daemon log: tail -f /var/log/lotus/daemon.log"
 echo "genesis daemon log:   tail -f /var/log/lotus/genesis-daemon.log"
 echo "genesis miner log:    tail -f /var/log/lotus/genesis-miner.log"
